+++ hexagonit/portletstyle/tests/base.py
@@ -8,7 +8,7 @@ from plone.app.testing import TEST_USER_ID
 from plone.app.testing import setRoles
 from plone.testing import z2
 
-import unittest2 as unittest
+import unittest
 
 
 class HexagonitPortletstyleLayer(PloneSandboxLayer):