+++ hexagonit/portletstyle/tests/test_controlpanel.py
@@ -12,8 +12,6 @@ from plone.registry.interfaces import IRegistry
 from zope.component import getMultiAdapter
 from zope.component import getUtility
 
-import unittest2 as unittest
-
 
 class TestControlPanel(IntegrationTestCase):
     """Integration tests for Portlet Styles control panel configlet."""
@@ -43,8 +41,7 @@ class TestControlPanel(IntegrationTestCase):
     def test_portlet_styles_in_controlpanel(self):
         """Check that there is an portletstyle entry in the control panel."""
         self.controlpanel = getToolByName(self.portal, "portal_controlpanel")
-        self.failUnless('portletstyles' in [a.getAction(self)['id']
-                            for a in self.controlpanel.listActions()])
+        self.failUnless('portletstyles' in [a.getAction(self)['id'] for a in self.controlpanel.listActions()])
 
     def test_record_portlet_styles(self):
         """Test that the portlet_styles record is in the control panel."""
@@ -55,9 +52,3 @@ class TestControlPanel(IntegrationTestCase):
             styles.portlet_styles,
             ['noheader|No header', 'nofooter|No footer', 'noheader nofooter|No header and no footer']
         )
-
-
-def test_suite():
-    """This sets up a test suite that actually runs the tests in the class
-    above."""
-    return unittest.defaultTestLoader.loadTestsFromName(__name__)