+++ hexagonit/portletstyle/tests/test_docstring.py
@@ -1,13 +1,11 @@
 # -*- coding: utf-8 -*-
 """Boilerplate for unit tests in method docstrings."""
 
-import unittest2 as unittest
+import unittest
 import doctest
 
 
 def test_suite():
     import hexagonit.portletstyle
 
-    return unittest.TestSuite([
-        doctest.DocTestSuite(hexagonit.portletstyle),
-        ])
+    return unittest.TestSuite([doctest.DocTestSuite(hexagonit.portletstyle)])